requirements.txt
pyyaml
colorlog
vosk
sounddevice
numpy
keyboard
pystray
Pillow
google-genai
